No political discussions
I haven't seen a single political discussion end well. There always seems to be a party that'll disagree. Personally, I find that it's just putting emphasis on situations that are already morally delicate to deal with and completely unrelated with computer science. Not that we should blind ourselves, but there are better places for that kind of topics.